backend/requirements.txt
flask==3.0.3
tensorflow-cpu==2.21.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
pycryptodome==3.20.0
pillow==10.3.0
werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
